# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.SyncTargetType:
    enum:
    - user
    - group
    - account
    type: string
    x-enum-comments:
      SyncTargetTypeAccount: 云账号级别同步
    x-enum-descriptions:
    - ""
    - ""
    - 云账号级别同步
    x-enum-varnames:
    - SyncTargetTypeUser
    - SyncTargetTypeGroup
    - SyncTargetTypeAccount
  domain.SyncTaskType:
    enum:
    - user_sync
    - permission_sync
    - group_sync
    - batch_user_sync
    type: string
    x-enum-comments:
      SyncTaskTypeBatchUserSync: 批量用户同步（从云平台拉取所有用户）
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - 批量用户同步（从云平台拉取所有用户）
    x-enum-varnames:
    - SyncTaskTypeUserSync
    - SyncTaskTypePermissionSync
    - SyncTaskTypeGroupSync
    - SyncTaskTypeBatchUserSync
...
  ginx.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider:
    enum:
    - aliyun
    - aws
    - azure
    - tencent
    - huawei
    - volcano
    - volcengine
    type: string
    x-enum-comments:
      CloudProviderAWS: Amazon Web Services
      CloudProviderAliyun: 阿里云
      CloudProviderAzure: Microsoft Azure
      CloudProviderHuawei: 华为云
      CloudProviderTencent: 腾讯云
      CloudProviderVolcano: 火山云 (别名)
      CloudProviderVolcengine: 火山引擎
    x-enum-descriptions:
    - 阿里云
    - Amazon Web Services
    - Microsoft Azure
    - 腾讯云
    - 华为云
    - 火山云 (别名)
    - 火山引擎
    x-enum-varnames:
    - CloudProviderAliyun
    - CloudProviderAWS
    - CloudProviderAzure
    - CloudProviderTencent
    - CloudProviderHuawei
    - CloudProviderVolcano
    - CloudProviderVolcengine
...
  github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTask:
    properties:
      cloud_account_id:
        type: integer
      create_time:
        type: string
      ctime:
        type: integer
      end_time:
        type: string
      error_message:
        type: string
      id:
        type: integer
      max_retries:
        type: integer
      progress:
        type: integer
      provider:
        $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider'
      retry_count:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTaskStatus'
      target_id:
        type: integer
      target_type:
        $ref: '#/definitions/domain.SyncTargetType'
      task_type:
        $ref: '#/definitions/domain.SyncTaskType'
      update_time:
        type: string
      utime:
        type: integer
    type: object
  github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTaskStatus:
    enum:
    - pending
    - running
    - success
    - failed
    - retrying
    type: string
    x-enum-varnames:
    - SyncTaskStatusPending
    - SyncTaskStatusRunning
    - SyncTaskStatusSuccess
    - SyncTaskStatusFailed
    - SyncTaskStatusRetrying
  service.GroupSyncResult:
    properties:
      created_groups:
        type: integer
      failed_groups:
        type: integer
      failed_members:
        type: integer
      synced_members:
        type: integer
      total_groups:
        type: integer
      total_members:
        type: integer
      updated_groups:
        type: integer
    type: object
...
  web.CreateSyncTaskVO:
    properties:
      cloud_account_id:
        type: integer
      provider:
        $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.CloudProvider'
      target_id:
        type: integer
      target_type:
        $ref: '#/definitions/domain.SyncTargetType'
      task_type:
        $ref: '#/definitions/domain.SyncTaskType'
    required:
    - cloud_account_id
    - provider
    - target_id
    - target_type
    - task_type
    type: object
...
  web.PageResult:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  web.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  web.SyncAccountReq:
    description: 同步云账号资产的请求参数
    properties:
      asset_types:
        description: |-
          要同步的资源类型列表，支持: ecs, rds, oss, vpc 等
          为空则默认只同步 ecs
        example:
        - ecs
        - rds
        items:
          type: string
        type: array
      regions:
        description: 要同步的地域列表，为空则同步账号配置的所有地域
        example:
        - cn-hangzhou
        - cn-shanghai
        items:
          type: string
        type: array
    type: object
  web.SyncAssetsReq:
    properties:
      account_id:
        description: 云账号ID
        type: integer
      asset_types:
        description: 要同步的资源类型，为空则同步所有支持的类型
        items:
          type: string
        type: array
    required:
    - account_id
    type: object
  web.SyncResult:
    description: 云资产同步操作的结果
    properties:
      message:
        description: 同步结果消息
        example: 同步完成，共同步 10 个资产
        type: string
      start_time:
        description: 同步开始时间
        type: string
      status:
        description: '同步状态: running, success, failed'
        example: success
        type: string
      sync_id:
        description: '同步任务ID，格式: sync_{account_id}_{timestamp}'
        example: sync_1_1706000000
        type: string
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业云资产管理服务 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: E-CAM Service API
  version: "1.0"
paths:
...
  /api/v1/cam/iam/groups/sync:
    post:
      description: |-
        从云平台同步用户组信息，包括用户组基本信息、权限策略和用户组成员。支持阿里云、腾讯云等多云平台。
        同步结果包含：用户组统计（总数、新建、更新、失败）和成员统计（总数、同步成功、失败）
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 云账号ID
        in: query
        name: cloud_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 同步结果统计
          schema:
            allOf:
            - $ref: '#/definitions/web.Result'
            - properties:
                data:
                  $ref: '#/definitions/service.GroupSyncResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/web.Result'
        "500":
          description: 同步失败
          schema:
            $ref: '#/definitions/web.Result'
      summary: 同步云平台用户组及其成员
      tags:
      - 用户组管理
  /api/v1/cam/iam/permissions/groups/{group_id}:
    get:
      parameters:
      - description: 用户组ID
        in: path
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户组的所有权限
      tags:
      - 权限管理
  /api/v1/cam/iam/permissions/policies:
    get:
      parameters:
      - description: 云账号ID
        in: query
        name: cloud_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 按云厂商查询可用的权限策略
      tags:
      - 权限管理
  /api/v1/cam/iam/permissions/users/{user_id}:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户的所有权限
      tags:
      - 权限管理
  /api/v1/cam/iam/permissions/users/{user_id}/effective:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取用户的有效权限（包含用户组继承的权限）
      tags:
      - 权限管理
  /api/v1/cam/iam/sync/tasks:
    get:
      parameters:
      - description: 任务类型
        in: query
        name: task_type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 云账号ID
        in: query
        name: cloud_account_id
        type: integer
      - description: 云厂商
        in: query
        name: provider
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.PageResult'
      summary: 查询同步任务列表
      tags:
      - 同步任务管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建同步任务请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CreateSyncTaskVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 创建同步任务
      tags:
      - 同步任务管理
  /api/v1/cam/iam/sync/tasks/{id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 获取同步任务状态
      tags:
      - 同步任务管理
  /api/v1/cam/iam/sync/tasks/{id}/retry:
    post:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Result'
      summary: 重试失败的同步任务
      tags:
      - 同步任务管理
...
  /api/v1/cam/iam/users/sync:
    post:
      description: 创建异步同步任务，立即返回任务ID，可通过任务ID查询同步进度
      parameters:
      - description: 租户ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: 云账号ID
        in: query
        name: cloud_account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Result'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Havens-blog_e-cam-service_internal_shared_domain.SyncTask'
              type: object
      summary: 同步云平台用户（异步）
      tags:
      - 用户管理
...
  /cam/assets/sync:
    post:
      consumes:
      - application/json
      deprecated: true
      description: 同步指定云账号的资产状态和信息。此接口已废弃，请使用 POST /api/v1/cam/cloud-accounts/{id}/sync
      parameters:
      - description: 同步资产请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.SyncAssetsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/ginx.Result'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/ginx.Result'
        "404":
          description: 云账号不存在
          schema:
            $ref: '#/definitions/ginx.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 同步云资产（已废弃）
      tags:
      - 资产发现
...
  /cam/cloud-accounts/{id}/sync:
    post:
      consumes:
      - application/json
      description: |-
        同步指定云账号下的云资产到本地数据库。这是同步云资产的主要接口，支持按资源类型和地域过滤同步范围。
        同步过程会自动获取云账号配置的所有地域，并逐个地域同步指定类型的资产。
        如果不指定 asset_types，默认只同步 ECS 实例。
      parameters:
      - description: 云账号ID
        in: path
        name: id
        required: true
        type: integer
      - description: 同步请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.SyncAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: 同步成功
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Result'
            - properties:
                data:
                  $ref: '#/definitions/web.SyncResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/ginx.Result'
        "404":
          description: 云账号不存在
          schema:
            $ref: '#/definitions/ginx.Result'
        "409":
          description: 云账号已禁用
          schema:
            $ref: '#/definitions/ginx.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 同步云账号资产（推荐）
      tags:
      - 云账号操作
...
schemes:
- http
- https
swagger: "2.0"